enhance popover view
