+++ Susi/Assets/Assets.xcassets/Logout.imageset/Contents.json
