+++ Susi/Assets/Assets.xcassets/Logout.imageset/ic_settings_power_2x.png
